aprendendo a usar tratamento de erro 'exceptions'

+++ src/main/java/com/roberto/Livro_de_Receitas/service/ReceitasService.java
@@ -6,12 +6,14 @@
 package com.roberto.Livro_de_Receitas.service;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+import com.roberto.Livro_de_Receitas.exception.RecursoNaoEncontradoException;
 import com.roberto.Livro_de_Receitas.model.ReceitasDB;
 import com.roberto.Livro_de_Receitas.repository.ReceitasRepository;
+
+import ch.qos.logback.core.joran.conditional.IfAction;
 
 @Service
 public class ReceitasService {
@@ -28,8 +30,10 @@ public class ReceitasService {
     }
 
     //CLASSE QUE BUSCA OS DADOS POR ID
-    public Optional<ReceitasDB> buscarPorId(Long id){
-        return receitasRepository.findById(id);
+    public ReceitasDB buscarPorId(Long id){
+        return receitasRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontradoException
+                ("Receita com ID "+id+" não encontrada!")); //CRIO A EXCEÇÃO SE NÃO ENCONTRAR A RECEITA
     }
 
     //CLASSE PARA SALVAR AS RECEITAS
@@ -39,6 +43,11 @@ public class ReceitasService {
 
     //CLASSE PARA DELETAR UMA RECEITA
     public void deletarReceita(Long id){
+
+        if(!receitasRepository.existsById(id)){
+            throw new RecursoNaoEncontradoException("Receita com o ID "+id+" não encontrado!");
+        }
+
         receitasRepository.deleteById(id);
     }
